+++ src/api/entity.ts
@@ -1,19 +1,13 @@
-import { utils, Wallet, Signer, ContractTransaction } from "ethers"
+import { Wallet, Signer, ContractTransaction } from "ethers"
 import { checkValidEntityMetadata, EntityMetadata } from "../models/entity"
-import { Gateway, IGateway, Web3Gateway, IWeb3Gateway } from "../net/gateway"
+import { Gateway, IGateway } from "../net/gateway"
 import { TextRecordKeys } from "../models/entity"
 import { fetchFileString, addFile } from "./file"
 import { IGatewayPool, GatewayPool } from "../net/gateway-pool"
 import { XDAI_CHAIN_ID, XDAI_GAS_PRICE, SOKOL_CHAIN_ID, SOKOL_GAS_PRICE } from "../constants"
-import { IMethodOverrides } from "dvote-solidity"
+import { IMethodOverrides, ensHashAddress } from "dvote-solidity"
 
-/**
- * Computes the ID of an entity given its address
- * @param entityAddress
- */
-export function getEntityId(entityAddress: string): string {
-    return utils.keccak256(entityAddress)
-}
+export { ensHashAddress } from "dvote-solidity"
 
 /**
  * Asserts that the given object is a valid Entity JSON metadtata.
@@ -25,17 +19,17 @@ export function checkValidMetadata(entityMetadata: EntityMetadata) {
 
 /**
  * Fetch the JSON metadata file for the given entity ID using the given gateway instances
- * @param entityId
+ * @param address
  * @param web3Gateway Web3Gateway instance already connected
  * @param dboteGateway Gateway instance already connected to an active service
  */
-export async function getEntityMetadata(entityId: string, gateway: Gateway | IGatewayPool): Promise<EntityMetadata> {
-    if (!entityId) return Promise.reject(new Error("Invalid entityAddress"))
+export async function getEntityMetadata(address: string, gateway: Gateway | IGatewayPool): Promise<EntityMetadata> {
+    if (!address) return Promise.reject(new Error("Invalid address"))
     else if (!(gateway instanceof Gateway || gateway instanceof GatewayPool)) return Promise.reject(new Error("Invalid Gateway object"))
 
     const resolverInstance = await gateway.getEntityResolverInstance()
 
-    const metadataContentUri = await resolverInstance.text(entityId, TextRecordKeys.JSON_METADATA_CONTENT_URI)
+    const metadataContentUri = await resolverInstance.text(ensHashAddress(address), TextRecordKeys.JSON_METADATA_CONTENT_URI)
     if (!metadataContentUri) return Promise.reject(new Error("The given entity has no metadata defined yet"))
 
     const jsonBuffer = await fetchFileString(metadataContentUri, gateway)
@@ -44,33 +38,19 @@ export async function getEntityMetadata(entityId: string, gateway: Gateway | IGa
 }
 
 /**
- * Fetch the JSON metadata file for the given entity address using the given gateway instances
- * @param entityAddress
- * @param web3Gateway Web3Gateway instance already connected
- * @param dboteGateway Gateway instance already connected to an active service
- */
-export function getEntityMetadataByAddress(entityAddress: string, gateway: Gateway | IGatewayPool): Promise<EntityMetadata> {
-    if (!entityAddress) return Promise.reject(new Error("Invalid entityAddress"))
-
-    const entityId = getEntityId(entityAddress)
-    return getEntityMetadata(entityId, gateway)
-}
-
-/**
- * Update the ENS fields on the blockchain and upload the corresponding JSON metadata file to IPFS using a Gateway
+ * Update the ENS Text fields on the blockchain and upload the corresponding JSON metadata file to IPFS using a Gateway
  * Throws an Error if the schema does not match
  * NOTE: The JSON metadata may need a few minutes before it can be generally fetched from IPFS
  * @return A content URI with the IPFS origin
  */
-export async function updateEntity(entityAddress: string, entityMetadata: EntityMetadata,
-    walletOrSigner: Wallet | Signer, gateway: IGateway | IGatewayPool): Promise<string> {
-    if (!entityAddress) return Promise.reject(new Error("Invalid entityAddress"))
-    else if (!entityMetadata) return Promise.reject(new Error("Invalid Entity metadata"))
+export async function setMetadata(address: string, metadata: EntityMetadata, walletOrSigner: Wallet | Signer, gateway: IGateway | IGatewayPool): Promise<string> {
+    if (!address) return Promise.reject(new Error("Invalid address"))
+    else if (!metadata) return Promise.reject(new Error("Invalid Entity metadata"))
     else if (!(gateway instanceof Gateway || gateway instanceof GatewayPool)) return Promise.reject(new Error("Invalid Gateway object"))
 
     // throw if not valid
-    checkValidEntityMetadata(entityMetadata)
-    const strJsonMeta = JSON.stringify(entityMetadata)
+    checkValidEntityMetadata(metadata)
+    const strJsonMeta = JSON.stringify(metadata)
 
     if (walletOrSigner instanceof Wallet && !walletOrSigner.provider) {
         walletOrSigner = walletOrSigner.connect(gateway.getProvider())
@@ -81,26 +61,26 @@ export async function updateEntity(entityAddress: string, entityMetadata: Entity
     // Set the IPFS origin on the blockchain
     const resolverInstance = await gateway.getEntityResolverInstance(walletOrSigner)
 
-    const entityId = getEntityId(entityAddress)
+    const entityAddrHash = ensHashAddress(address)
     const chainId = await gateway.getChainId()
     let options: IMethodOverrides
-    let tx : ContractTransaction
+    let tx: ContractTransaction
     switch (chainId) {
-        case XDAI_CHAIN_ID :
+        case XDAI_CHAIN_ID:
             options = { gasPrice: XDAI_GAS_PRICE }
-            tx = await resolverInstance.setText(entityId, TextRecordKeys.JSON_METADATA_CONTENT_URI, ipfsUri, options)
+            tx = await resolverInstance.setText(entityAddrHash, TextRecordKeys.JSON_METADATA_CONTENT_URI, ipfsUri, options)
             break
-        case SOKOL_CHAIN_ID :
+        case SOKOL_CHAIN_ID:
             const addr = await walletOrSigner.getAddress()
             const nonce = await walletOrSigner.provider.getTransactionCount(addr)
             options = {
                 gasPrice: SOKOL_GAS_PRICE,
                 nonce,
             }
-            tx = await resolverInstance.setText(entityId, TextRecordKeys.JSON_METADATA_CONTENT_URI, ipfsUri, options)
+            tx = await resolverInstance.setText(entityAddrHash, TextRecordKeys.JSON_METADATA_CONTENT_URI, ipfsUri, options)
             break
-        default :
-            tx = await resolverInstance.setText(entityId, TextRecordKeys.JSON_METADATA_CONTENT_URI, ipfsUri)
+        default:
+            tx = await resolverInstance.setText(entityAddrHash, TextRecordKeys.JSON_METADATA_CONTENT_URI, ipfsUri)
 
     }